+++ detectable/src/test/java/com/synopsys/integration/detectable/detectables/bazel/unit/BazelDetectableTest.java
@@ -24,7 +24,7 @@ public class BazelDetectableTest {
         Mockito.when(fileFinder.findFile(new File("."), "WORKSPACE")).thenReturn(new File("src/test/resources/functional/bazel/WORKSPACE"));
         BazelExtractor bazelExtractor = null;
         BazelResolver bazelResolver = null;
-        BazelDetectableOptions bazelDetectableOptions = new BazelDetectableOptions("target", null, null, null);
+        BazelDetectableOptions bazelDetectableOptions = new BazelDetectableOptions("target", null, null);
         BazelDetectable detectable = new BazelDetectable(environment, fileFinder, bazelExtractor, bazelResolver, bazelDetectableOptions.getTargetName().orElse(null));
 
         assertTrue(detectable.applicable().getPassed());